Functions

SemiAlgebraicTypes.mesh — Function

Mesh of the parametric curve f: u -> [x(u),y(u),z(u)] for u in the interval U.

c = mesh(u->[u,sin(u^2),cos(2*u)], 0.0 => 2.0*pi, 1000; field=DistField(0.0,0.0,0.0))

Mesh of the parametric surface f: (u,v) -> [x,y,z] for u in the interval U, v in the interval V. Example

mesh((u,v)->[u,v,cos(2*u*v)], 0.0 => 2.0, -pi => pi, field=DistField(0.0,0.0,0.0))

SemiAlgebraicTypes.sample — Function

Sample the parametric curve f: u -> [x,y,z] for u in the interval X.

U = sample(u->[u,sin(u^2),cos(2*u)], 0.0 => 2.0*pi, 1000)

Matrix of points of the parametric surface f: (u,v) -> [x(u,v),y(u,v),z(u,v)] for u in the interval U, v in the interval V. Example

sample((u,v)->[u,v,cos(2*u*v)], 0.0 => 2.0, -pi => pi)

SemiAlgebraicTypes.polar — Function

Mesh of the graph of the radius function r = f(x,y,z) on the unitary sphere.

polar((x,y,z)->cos(x*y*z+1.0), 200, field=DistField())

SemiAlgebraicTypes.box — Function

box(c::Vector{T}, r::Vector)

Compute the mesh corresponding to a box aligned with the axes and centered at the point c of size 2r.

SemiAlgebraicTypes.cube — Function

cube(c::Vector{T}, r::T)

Compute the mesh corresponding to a cube aligned with the axes and centered at the point c of size 2r.

cube(p1::Vector{T}, p2::Vector{T})

Compute the mesh corresponding to a cube aligned with the axes with diagonal points p1, p2.