Mesh

SemiAlgebraicTypes.Mesh — Type

Mesh{T}

Mesh corresponding to an array of points (Vector{T}), an array of edges (Vector{Int64}) and an array of faces (Vector{Int64}).

Example

julia> mesh(Float64);

julia> mesh([[cos(i*pi/5), sin(i*pi/5), 0.0] for i in 1:10], Edge[], [[1,i,i+1] for i in 1:9]);

Fields:

points ::Matrix{T}: array of points

edges ::Vector{Vector{Int64}}: array of edges

faces ::Vector{Vector{Int64}}: array of faces

normals ::Matrix{T}: array of normals

attr ::Dict{Symbol,Any}: attributes

Accessors and modificators

SemiAlgebraicTypes.push_vertex! — Function

Insert a vertex at the end of the vertex array of a mesh.

julia> m = mesh(Float64,3);

julia> push_vertex!(m,[1.,2.,3.])
SemiAlgebraicTypes.Mesh{Float64}(Array{Float64,1}[[1.0, 2.0, 3.0]], Array{Int64,1}[], Array{Int64,1}[], Dict{Symbol,Any}())

SemiAlgebraicTypes.push_edge! — Function

Insert a new edge given by the array of indices of the points (numbering starting at 1) at the end of the edge array of the mesh.

julia> m = mesh(Float64);

julia> push_vertex!(m,point(0.,0.,0.)); push_vertex!(m,point(1.,0.,0.)); push_edge!(m,[1,2])
SemiAlgebraicTypes.Mesh{Float64}(Array{Float64,1}[[0.0, 0.0, 0.0], [1.0, 0.0, 0.0]], Array{Int64,1}[[1, 2]], Array{Int64,1}[], Dict{Symbol,Any}())

SemiAlgebraicTypes.push_face! — Function

Insert a new face with the array of indices of the points (numbering starting at 1) at the end of the face array.

julia> m = mesh(Float64);

julia> for i in 1:10 push_vertex!(m,[cos(i*pi/5), sin(i*pi/5), 0.0]) end;

julia> for i in 1:9 push_face!(m,[1,i,i+1]) end;

Constructors

SemiAlgebraicTypes.cube — Function

cube(c::Vector{T}, r::T)

Compute the mesh corresponding to a cube aligned with the axes and centered at the point c of size 2r.

cube(p1::Vector{T}, p2::Vector{T})

Compute the mesh corresponding to a cube aligned with the axes with diagonal points p1, p2.